+++ src/app/(auth)/auth-layout.tsx
@@ -1,0 +1,10 @@
+const AuthLayout = ({ children, title }: { children: React.ReactNode; title: string }) => {
+  return (
+    <section className="flex flex-col items-center justify-center font-pretendard">
+      <div className="my-12 text-2xl font-bold">{title}</div>
+      {children}
+    </section>
+  );
+};
+
+export default AuthLayout;
