+++ src/app/(auth)/login/page.tsx
@@ -1,69 +1,28 @@
-'use client';
+import GoogleButton from '@/components/common/Button/GoogleButton';
+import KakaoButton from '@/components/common/Button/KakaoButton';
+import Link from 'next/link';
+import AuthLayout from '../auth-layout';
+import LoginForm from '@/components/common/Form/LoginForm';
+import { Metadata } from 'next';
 
-import React from 'react';
-import { FieldValues, useForm } from 'react-hook-form';
-import InputField from '../../../components/common/InputFeild/InputField';
-import { useRouter } from 'next/navigation';
+export const metadata: Metadata = {
+  title: '로그인 페이지',
+  description: '로그인 페이지 입니다.'
+};
 
 const LoginPage = () => {
-  const router = useRouter();
-  const {
-    register,
-    handleSubmit,
-    formState: { errors }
-  } = useForm({
-    mode: 'onSubmit'
-  });
-
-  const onSubmit = async (userInfo: FieldValues) => {
-    const res = await fetch('/api/login', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        ...userInfo
-      })
-    });
-    const { user } = await res.json();
-
-    if (!user) {
-      alert('아이디와 비밀번호를 다시 입력해주세요.');
-      return;
-    }
-
-    router.push('/');
-  };
-
   return (
-    <form onSubmit={handleSubmit(onSubmit)}>
-      <div>
-        <InputField
-          name="email"
-          type="email"
-          placeholder="이메일를 입력해 주세요."
-          registerOptions={register('email', {
-            required: '이메일을 다시 입력해주세요.'
-          })}
-          errors={errors}
-        />
-        <InputField
-          name="password"
-          type="password"
-          placeholder="비밀번호를 입력해 주세요."
-          registerOptions={register('password', {
-            required: '비밀번호를 다시 입력해주세요.'
-          })}
-          errors={errors}
-        />
-      </div>
-      <button
-        type="submit"
-        className="mt-9 h-[50px] w-[348px] rounded-md bg-black font-bold text-white shadow-buttonShadow hover:bg-zinc-800"
-      >
-        로그인
-      </button>
-    </form>
+    <AuthLayout title="로그인">
+      <LoginForm />
+      <Link href={'/reset-password'} className="mb-12 mt-3 text-sm text-[#999] hover:underline">
+        비밀번호 찾기
+      </Link>
+      <GoogleButton />
+      <KakaoButton />
+      <Link href={'/signup'} className="mt-12 text-sm font-bold hover:underline">
+        회원가입 &gt;
+      </Link>
+    </AuthLayout>
   );
 };
